User Story 27: Shelters can be deleted as long as all pets do not have approved applications on them

+++ app/controllers/shelters_controller.rb
@@ -27,7 +27,10 @@ class SheltersController < ApplicationController
 
   def destroy
     shelter = Shelter.find(params[:id])
-    shelter.destroy
+    if shelter.has_all_adoptable_pets?
+      shelter.delete_pets
+      shelter.destroy
+    end
     redirect_to '/shelters'
   end
